refactor: Transaction history comppnent
